Remove jQuery dependency
These days, JavaScript supports most of the features natively.
So this will make the extension lighter.

The exception is the :visible selector, which had to be copied from jQuery.

This is mostly one-to-one conversion, there is still space for modernizing the code
and making it more idiomatic (e.g. replacing loops with functional constructs).