Port to ES modules
We want to use a library for generating strings from regular expressions
in pattern attribute but both of the top contenders (randexp and reregexp)
are distributed as CommonJS modules. Those set `exports` global variable
so just executing them as content scripts is not sufficient
to get the library functions into the scope.

We could create a content script initializing the variable
`window.exports = {};` to run before each third-party library
and another one like `const { Chance } = window.exports;`
to be executed after but that would be very ugly.

Instead, I have decide to employ a module loader, which is cleaner.
But Firefox only supports ES modules so we need to emulate CommonJS’s
`require` function using that. And `import`ing only works in ES modules
so the file that requires the library will need to be a module. And
likewise will the files depending on the module, ad infinitum.
And `chrome.tabs.executeScript` will not run files as modules
with Manifest before V3 so we need to hack around it using `code` with `import`.

Alternately we could start using a bundler tool such as Parcel, which
would handle CommonJS loading just fine, producing a single file with
all libraries inserted in the correct order.
But I have decided not to go that way since it adds too much complexity
and makes feedback loop slower during development, especially when
the way hot module reloading is implemented will not work in Chrome.

Additionally, explicit imports make it cleaner where code comes from,
which is a benefit for code understandability.